getDashboardBusinesses and createAddresses
Change address logic and get the business by the distances and the delivery cost was calculated

+++ app/Http/Controllers/BusinessController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Address;
 use App\Models\Business;
 use Facade\FlareClient\Http\Response;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Bus;
 
 class BusinessController extends Controller
 {
@@ -42,12 +44,87 @@ class BusinessController extends Controller
     }
 
 
+    //this method get the businesses by category
     public function getBusinessesByCategory($categoryId)
     {
         $businesses = Business::whereCategoryId($categoryId)->get();
-        return Response()->json($businesses);
+
+        foreach ($businesses as $business) {
+        }
+
+        return response()->json($businesses);
     }
 
+
+    public function getDashboardBusinesses(Request $request)
+    {
+
+        //Get businesses by closer location and better rating  || Sorting || Ads
+
+
+        $latClient = $request->latClient;
+        $lngClient = $request->lngClient;
+        $businesses = Business::all();
+        $businessesArray[] = $businesses;
+        $businessesArrayResponse = [];
+
+        foreach ($businesses as $business) {
+
+            $businessAddress = Address::whereBusinessId($business->id)->first();
+            $distance = $this->calculateDistance($businessAddress->lat, $businessAddress->lng, $latClient, $lngClient);
+            switch ($distance) {
+                case $distance <= 3:
+                    $business['deliveryCost'] = 25;
+                    break;
+                case $distance <= 6:
+                    $business['deliveryCost'] = 35;
+                    break;
+                case $distance <= 10:
+                    $business['deliveryCost'] = 45;
+                    break;
+                case $distance <= 15:
+                    $business['deliveryCost'] = 50;
+                    break;
+                default;
+                    $business['deliveryCost'] = 50;
+                    break;
+            }
+            $business['distance'] = $distance;
+            $businessesArrayResponse[] = $business;
+        }
+        //sort businesses by distance
+        usort($businessesArrayResponse, array($this, 'sort_businesses_by_distance'));
+        return response()->json($businessesArrayResponse);
+    }
+
+
+    private static function sort_businesses_by_distance($a, $b)
+    {
+        if ($a->distance == $b->distance) {
+            return 0;
+        }
+        return ($a->distance < $b->distance) ? -1 : 1;
+    }
+
+
+    public function calculateDistance($latFrom, $lngFrom, $latTo, $lngTo)
+    {
+
+        $earthRadius = 6371000;
+        // convert from degrees to radians
+        $latFromConverted = deg2rad((float)$latFrom);
+        $lonFromConverted = deg2rad((float)$lngFrom);
+        $latToConverted = deg2rad((float)$latTo);
+        $lonToConverted = deg2rad((float)$lngTo);
+        $latDelta = $latToConverted - $latFromConverted;
+        $lonDelta = $lonToConverted - $lonFromConverted;
+        $angle = 2 * asin(sqrt(pow(sin($latDelta / 2), 2) +
+            cos($latFrom) * cos($latTo) * pow(sin($lonDelta / 2), 2)));
+        //distance in meters
+        $distance = ($angle * $earthRadius) / 1000;
+
+        return round($distance, 3);
+    }
 
     public function uploadPhoto(Request $request, $business)
     {
@@ -64,8 +141,11 @@ class BusinessController extends Controller
             $object = $bucket->upload($uploadedfile, ['name' => $firebase_storage_path . $file, 'predefinedAcl' => 'publicRead']);
             $publicUrl = "https://{$bucket->name()}.storage.googleapis.com/{$object->name()}";
             //will remove from local laravel folder  
-            unlink($localfolder . $file);
-
+            if (file_exists($localfolder . $file)) {
+                   unlink($localfolder . $file);
+                } else {
+                   echo "The file filename does not exist";
+                }
             return $publicUrl;
         } else {
             echo 'error';
